tools: improve uniform value printing
Memory address to buffers can be passed via uniforms to the shaders. To make RE a
little bit easier try to check if the value is a known allocated memory address and
change the printed output to show the name of the buffer and its address value in hex.

before:
    0x00000001, /*   [05000] VS.UNIFORMS[0] := 0.000000 (u0.x) */
    0x63766e00, /*   [05004] VS.UNIFORMS[1] := 4545825377476721770496.000000 (u0.y) */

after:
    0x00000001, /*   [05000] VS.UNIFORMS[0] := 0.000000 (u0.x) */
    0x63766e00, /*   [05004] VS.UNIFORMS[1] := ADDR_VERTEX_27 (0x63766e00) (u0.y) */

Signed-off-by: Christian Gmeiner <cgmeiner@igalia.com>